frontend/pyproject.toml
[project]
name = "frontend"
version = "0.0.1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-storage-blob>=12.25.1",
    "python-dotenv>=1.1.0",
    "streamlit>=1.44.1",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages]
find = {}

[tool.setuptools.package-data]
"*" = ["*"]

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.25.0",
    "pytest-httpx>=0.24.0",
    "pytest-mock>=3.11.0",
    "black>=23.9.1",
    "isort>=5.12.0",
    "mypy>=1.5.2",
    "flake8>=6.1.0",
]
dev = [
    "python-dotenv>=1.0.0"
]

[tool.uv]
dev-dependencies = [
    "poethepoet>=0.29.9",
    "pre-commit>=4.1.0"
]
[tool.poe]
include = "poe.toml"

[tool.pyright]
venvPath = "."
venv = ".venv"

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version_scheme = "pep440"
version_provider = "pep621"
update_changelog_on_bump = true
major_version_zero = true
version_files = [
    "pyproject.toml",
    "docker/Dockerfile",
    "src/app.py",
]

bump_message = "chore(bump): $current_version -> $new_version [skip ci]"
